Search: Prioritize notes with search matches in their title
We have been returning search results in no particular order. When we
get searches that return many notes though the noise can make the search
less valuable.

In this patch we're starting to prioritize search results where the
first line of the note contains the string searched for. This is in
effect saying that we're prioritizing the note title matches in
searches.