Fixes markdown code styles (#1702)
We were importing the highlight.js styles but only using parts of them.
Solarized-light and dark were both imported but only dark was being used as the
styles were being made global. We covered this up by modifying the background
color and text color to match the theme. This removes those cover ups and uses
the scss so the resulting css is not scoped globally.