Crash on missing platform support (#1713)
The Simplenote app, for better or worse, currently depends on a few
storage technologies that may be disabled through browser settings,
extensions, and manual configuration.

When these dependencies are missing we cannot function normally and
right now the app crashes or breaks in misleading ways.

In this patch we perform an up-front check when the app boots to verify
that we have some basic dependencies met. This is not an exhaustive
validation but it should catch the most common cases and give customers
an idea of why things may be broken.