Fix: Stop infinite duplication by correcting typo (and stuff) (#1724)
Resolves #1690
Obviates session lock in: #1700, #1704, #1707, #1710, #1720
Resolves bug that uncovered signout/signin issue: #1664
Follows Simperium API change in: #1598, #1599, Simperium/node-simperium#61, Simperium/node-simperium#78

When we fixed some deep and underlying issues in `node-simperium` we
started a chain of operations which had to adjust to that change.
Remember that the core problem was an assumption that after we send out
a change that we could wait until it came back. That assumption was
wrong because changes from other remote clients could come in during
that time period while we're waiting. Simperium/node-simperium#61 fixed
that problem but because it was broken in the past we added some
work-arounds in #706; these work-arounds complicated the flow of data in
the corrected version so we had to remove them and change how we handle
remote updates in the Simplnote side.

A critical step in this flow is letting Simperium know what local
changes we may have in the app buffer. We introduced this step in #1598
but never noticed a typo:

```js
getState().selectedNoteId
```

That reference will always be wrong and so we'll never send the local
contents of the note. It should have been...

```js
getState().appState.selectedNoteId
```

The implication of this change is that we always told Simperium that
there are no local changes and so it would grab the currenty copy in the
bucket, which in our case is in shared `indexedDB` tables, and thus it
would occasionally get updated copies of the bucket that were updated in
another session. This would then appear as though we also made the same
change we were receiving, and thus we would prepare our own patch to
send out and start the cycle.

Further we were only sending local updates in the case that we had the
same note open as was updated in the remote change. However, since we
don't update Simperium with notes that we're not editing we were leaving
other notes behind by sending `null` for them. The outcome of this is
that edits to notes that aren't selected would have a modification in
the way they are applied, if otherwise they would have started the
infinite looping.

After this change we're always sending the current local copy. This is
still somewhat of a work-around but it's one that should reliably cover
over the larger problems in the system. Eventually we have to clear out
the data flow, remove local copies of our data, separate independent
clients or sessions so they don't share buckets, and simplify the
control flow to make things easier to debug.