Fixes bug where window is dragged when trying to select a note revision
There was a div that was draggable underneath the revision selectror.
Additionally the revision selector was draggable. In rememdying this we are now
longer rendering the note toolbar header while the revision panel is open. We
are also not rendering the revision panel when it is not open. The showing and
hiding of these panels is overly complex and should be looked at in a future
PR.