Fix settings URL so that it send the correct url param to GAE (#1736)
We were directing users to /settings?source=blah which GAE automatically
redirected to /settings/ removing the url param. Additionally, the code always
returned 'electron' as the source which is not only incorrect but GAE only
understands the source as react. From the web app and from electron we should
send the source as react so that we hide settings that only apply to the old
web app.