Fixes bug where window is dragged when trying to select a note revision (#1741)
There was a div that was draggable underneath the revision selector.
Additionally, the revision selector was draggable. In remedying this we are no
longer rendering the note toolbar header while the revision panel is open. We
are also not rendering the revision panel when it is not open. The showing and
hiding of these panels is overly complex and should be looked at in a future
PR.