Undo unintentional changes introduced in master
See #1754 - missed a few more changes that weren't intended

In this patch we're also adopting the `package.json` and
`package-lock.json` found in `develop` at the merge base of the
`release/1.12` branch. This means that for this commit in `master`
things will be a bit inaccurate because `master` never had this set of
dependencies, but it _will_ make the merge of `release/1.12` apply
without conflict and get us tracking again with the mainline
development.