Adds a Clipboard Button component (#1743)
The inline button component to copy the publish url was lacking an indication
that the text was copied. Additionally, the current setup was using a copy
method that can have an issue in some browsers. This commit switches to use of the
clipboard npm repo, moves code into a component, and adds an indication that
the text was copied.