Add TypeScript support and demo small refactor
As if the types and data flows weren't complicated enough in this
application, we have no support from the language or compiler to make
sense of it all.

In this patch we're adding TypeScript into the build pipeline to allow
for writing modules in TypeScript and to aid us in knowing just what we
are dealing with.

Types are stripped from the modules before sending to Babel and so
therefore we won't prevent a build just because there are type errors.

This patch enables a certain level of autocomplete on `appState`,
indicating which action creators are available.