App-state: Remove unused `indexCtr` (#1796)
In the earlier stages of the app we relied on counting index operations
to defer updating the internal note list. This was to prevent thrashing
in the app when receiving updates for an account with lots of notes.

In #89 we moved the `indexCtr` value out of the controlled app state and
turned it into an instance variable on the `actionMap` object. This
change was made because we couldn't wait for a Redux update cycle to
update that value - the instance property updated immediately and
imperatively.

In #987 we removed the index-counting operation entirely and instead
relied on a more natural `debounce()` on `bucket.on('index')`. This
removed the need for the work-around but didn't remove the `indexCtr`
property.

In this patch we're finally removing it to finish the work from #987.
The property is currently unused and unreferenced in the app.