Refactor: Remove "pinned" exceptions (#1795)
To understand this change we have to undestand some core data models in
the app and some of the app's history.

Most data structures are composed of two parts: the Simperium `Entity`
which includes Simperium information of id, version, and wrapped data;
and the app-specific `data` which comes from Simplenote. That is, a
`note` is composed of the `Note` type which includes things like
`content` and `tags` and the `NoteEntity` which is the wrapped version.

We store all of the Simperium entities in `indexedDB` and in the
inception of this app we told `indexedDB` to create a search index based
off of whether or not a note was pinned. This was to make it fast to
list the notes with pinned notes at the top.

In order to facilitate the `pinned` index we had to create a fake
top-level property called `pinned` becuase we can't tell the browser:
"build an index on whether the `"pinned"` value is in the
`data.systemTags` array."

However, in [229a76b] we removed this index and moved pinned-sorting
into the application itself. When we did that we left behind the relic
of this top-level `pinned` value.

In this patch we're removing that relic. Why? Because it creates another
level of complexity by special-casing the `pinned` value which we have
to know about and reference when looking at the pinned status of a note.
By undoing this exception we can rely on the more semantically-native
means to check if or set whether a note is pinned.

This patch was triggered by what appeared to be an error when
introducing TypeScript types. We were looking at `note.pinned` which
shouldn't exist on the Simperium `Entity<Note>`. It _wasn't_ an error in
the app because the value existed but it wasn't clear that it _should_
exist.

I have preferred to undo the exception rather than to adjust the type.

[229a76b]: 229a76b4e6992728c20746cb5b6d0d3d32d27571

Add release notes

Fix broken tests

They were using incomplete Note models and therefore failed
when the app was relying on completer models.

They are now completer