Filter notes in state vs. on `render()`
The `filterNotes()` function is relatively expensive, particularly for accounts
with many notes.  Previously we have been running this several times in a
normal app update because it gets called in `mapStateToProps` on `note-list`
and also in a couple of other components. When these all re-render we end up
filtering the notes multiple times in a row even though no changes could have
occurred during that window of time (due to the single-threaded nature of
JavaScript, not because of the unlikelihood of receiving remote changes).

In this patch we're separating out the filtered notes into a new place in
`state` and we're updating it via middleware that watches for any actions which
might impact the note list. This gives us more control over the update cycle
and lets us replace debounced logic with synchronous updates.

After this change the search field's input will immediately update to reflect
the changes being entered in on the keyboard or via the tag list panel but the
note list won't immediately update when only typing in the search field. It
will update immediately when changing tags or when performing "atomic" updates
like trashing or restoring a note. The decision to defer or operate
"immediately" was based on the expectations of how interactive each action
should be. When we're typing in a search field we are probably going to type
more before we need the results to update, but when we flip to another tag we
expect that to immediately reflect visually.