Add missing decorators in note list previews
When showing each item in the note list we run the content decorators on them
in order to show checkboxes instead of the markdown syntax for a checkbox and
we also highlight the matching text from the search box.

Previously we have only been showing the first match for each decorator because
of a couple small bugs:
 - the task prefix matcher wasn't using the `m` multi-line flag in its RegExp
   pattern
 - the search filter text was only replacing the first occurrance, see JS
   "replaceAll()" proposal

In this patch we're rewritten `decorateWith` to automatically create a global
RegExp pattern for the search text and the task prefix RegExp pattern has been
corrected.

After this patch you can now see all of the decorated items in the note list;
for example, if you had two checklist items in a note then you can now see
multiple checkboxes in the note list whereas before you'd only see the first
one and then the second one would appear as ` - [ ]` or `- [x]`