Fix typo from #1792 preventing filtering by tag (#1817)
In #1792 I introduced a typo from `name` to `namer` and this prevented the tag
drawer from working properly. That is, when trying to filter by tag it was
looking for matches on `tag.data.namer` instead of on `tag.data.name` and
unsuprisingly there were no matches.

After this patch things are restored to normal working order.