Fixes focus mode toggle
When the app loads there is potential for the `is-note-open` css class to be
missing on app-layout. This removes the reliance on isNoteOpen as it is note
needed and should be independent. Now focus mode class will remain on all
screen widths but will only change the layout for widths greater than 750px. We
should look at why `is-note-open` is not applied to app layout and how layout
is setup in general.