Fix ol numbering in markdown preview (#1823)
Fixes: #676

The numbering of ol lists is currently broken when you have a list that is
interrupted with a line of text that is not part of the list.

This PR whitelists the start attribute on ol tags which fixes the issue.

For example:

1. One
1. Two
1. Three
some_text some_text
1. Four
The above example should number the list 1 to 4. It, however, orders the list 1
to 3 and then after the interrupt starts over at 1.

The output from showdown, the markdown to html processor, is:

<ol>
<li>One</li>
<li>Two</li>
<li>Three</li>
</ol>
<p>some_text some_text</p>
<ol start="4">
<li>Four</li>
</ol>
Notice the second ol has a start attribute <ol start="4">.

When we pass the html into the sanitizeHtml function that attribute is
stripped. This PR whitelists that attribute so it is not stripped.