Fixes thrashing between notes
We have reports of users who has seen the web and desktop apps thrashing
between two notes. What happens is that the user will get into a state where
for some reason the selected note is repeatedly changing back and forth between
two notes.

I believe that this was happening because the componentDidUpdate on the note
list was setting the selected note id if the note id was different than
previous props. My suspicion is that it was putting the selected note id into
an infinite loop of changing back and forth between two notes.

I have not been able to replicate this bug but this seems like the proper
solution.