Refactor the tag suggestions filtering function
I've removed the RegExp and rewritten the filtering loop. Why?

The RegExp added a step of confusion where we needed to parse and understand
what it was doing. It turns out that in the course of this refactor I found a
case where it wasn't doing what we thought it was doing, skipping a full-text
match for `tag:` given a tag named `tag:`. Without the RegExp we simply have
fewer concepts squished into this function.

The filtering loop also was iterating over every tag in the account and
checking if they matched the search query before it thought about pruning down
the list of matches to the maximum number. With the new loop we're using a
helper function (to keep the looping details out of the logic) which will abort
out of its looping as soon as we find as many results as we'll allow. Otherwise
if it never finds the maximum results it will continue iterating until it has
scanned every tag.

One final bug was fixed: it was previously possible to enter whitespace into
the search field and then it would look for all of the tags as possible
matches. With only whitespace in the search field there should be no tag
suggestions.