Search: Offload processing to WebWorker and free up main thread
As long as our `NoteList` has been bound up in re-computing the fitlered notes on every render and on every call to `mapStateToProps` and as long as it has been calculating note indices based on that list it calculates we have been in a pickle when it comes to rendering performance whil searching.

We have taken a few steps towards mitigating the CPU bottleneck:
 - debounced the search filter
 - debounced searching when updating the filter
 - moved the filtered notes and tags into Redux and only
   compute them when a debounce times out after events change
   the filtering: e.g. a search or a remote update

In this patch we're doing what we've long waited to do which is move the search filtering out of the main thread and into a WebWorker. By doing this we not only eliminate one of the biggest performance bottlenecks for accounts with lots of data but we have opened ourselves up to dramatically increasing the quality of our search results since the costs of performing the searching and indexing can be overlooked as the main UI thread remains responsive.

This specific patch only takes the first step which is reproducing the search as-is. It also trades some of the performance gain by passing all of the notes and tags in the account to the WebWorker on every update. However, as this is a relatively fast operation it shouldn't pose any significant obstacle.

If we develop this design for search indexing one area we can start improving is to stop sending the full contents of all the notes and tags on every fitlering event. It's enough to pass the search filter, whether the trash is open, and if there's a selected tag. We can do this by trapping more actions in the middleware and maintaining a copy of the notes in the WebWorker.

We can probably do even better by maintaining our copy of the notes independently, by pulling from `indexedDB`, though that may carry with it the higher risk of further coupling our Ghost and Bucket implementations to a shared-state system we are trying to avoid.

Once the data-passing and synchronization is figured out we can start to perform more robust indexing of the notes and rank search results.