Refactor: Use direct comparison of note id when receiving remote upates (#1838)
In `onNoteBeforeRemoteUpdate` we give ourselves a chance to rebase local edits
on a note before remote changes are applied. Although we haven't identified any
problem related to this refactor, we have been using indirect measures to check
if the note we are working with is the note we think it is. That is, we were
checking against `selectedNoteId` as a proxy for `note.id` when relying on
`note.id` works and eliminates the possibility that `selectedNoteId` doesn't
match `note.id`

In this patch we're simplifying to rely solely on that direct measure and
eliminting one possible bug.