Types: Add legacy action types (#1873)
When @belcherj asked about creating types for the existing actions in
`app-state` I was reluctant since we want to get rid of them but then in
considering his view I think it's the right decision until they are all gone;
so I have created the types here.

These are organized as one big `LegacyAction` type that isn't exported. This
provides auto-completion and type-checking while still hiding the types
somewhat. They are at the end of `action-types.ts` so that we can mostly ignore
them. As we remove functionality from `app-state` we should remove the
corresponding legacy action types.