Fix: Actually reset auth when authorizing without acess token
Discovered via TypeScript in #1884

Previously when authorizing an account which yields no `access_token`
we have been attempting to reset the authorization state by dispatching
`resetAuth`. Unfortunately we haven't been dispatching the function
call but rather the function/action-creator itself leading to no
actionable dispatch.

In this patch we're fixing the call by calling `resetAuth()` which
then passes a real action to the dispatcher.

This should fix an unsurfaced but during invalid authentication flows.