Refactor getNoteTitles to eliminate errors
This PR refactors getNoteTiles to make the code a little cleaner, add types to
function, and removes lodash dependency.

When you have 0 notes and have unsynced changes you can cause the function to
crash the app. In my testing, I was not able to create these test conditions.
You will see in the modified code that we ensure that we at very least have an
empty array before attempting array functions. This eliminates the existing
bug.