Refactor: Replace search field focus state with middleware
Alternative idea to #1908

`searchFocus` isn't quite app state. We have been using it as a
workaround in order to trigger side-effects from Redux in order
to focus the search field.

In this patch we're removing that state altogether and replacing
it with a custom middleware specific to the search field. This too
is somewhat ugly because of the coupling between the middleware
and the search field component; it needs more thought on how best
to register the DOM node containing focus.

Nevertheless by removing the `state` property we are able to remove
more code and the update loop that was necessary to set focus,
update, an then record that the focus was already set or unset.

Storing a state value for focus doesn't match the reality that the
focus is a property of the DOM node and changes independently from
the Redux actions being dispatched in the app.