Properly wait for DOM nodes to appear
This patch fixes two bugs related to `waitForExist()` that led to the
tests passing when they should have failed.

The first bug was a type error whereas `waitForExist()` returns a
promise-like object and `expect(...waitForExist()).toBeTruthy()`
would therefore always be true. This meant that we not only didn't
wait for the elements to appear but also that our check never failed.

This has been fixed by awaiting the return. This has been handled
inside of a new test helper function to cut down on incidental noise.

The second bug was that, despite the type signaures in the webdriverio
type definitions, `app.client` and `app.client.$()` return different
types, whereas `app.client` returns a `Client` and `$()` returns a
"WebElement JSON Object" or something that _isn't_ the same as the docs
suggest.

This bug was an infinite hang when trying to wait on the elements to
appear since the returned value never resolved.

Now all calls to wait on an element are handled by
`app.client.waitForExist(selector, ...)` instead of using `$()`