Fix: Render unicode bullet as list in Markdown preview (#1922)
Previously we have been failing to render bulleted-lists which use the
Unicode `\u2022` "Bullet". The preview would fail to show the line as
a new list item and therefore inline it and any successive list items
into a single paragraph/item.

Now we are replacing those bullets before rendering the Markdown so
that it will render properly as any other bullet would.