e2e tests: improve reliability of login test (#1913)
Previously the tests were a bit flakey, working some times and not others.  In
this patch we've added the `--runInBand` option to `jest` so that the tests run
sequentially in the same process as the test runny. It may have been that
`jest` was attempting to run the tests in parallel even though the nature of
testing Electron/Spectron doesn't support this and so we would sometimes fail
based on concurrency conflicts.

A few new tests have been added in order to verify that the new updates are
working and to demonstrate how we can start adding testable behaviors into the
app. We still need to improve how the tests are organized and setup to minimize
the noise in the code and isolate the tests.

This patch fixes two bugs related to `waitForExist()` that led to the
tests passing when they should have failed.

The first bug was a type error whereas `waitForExist()` returns a
promise-like object and `expect(...waitForExist()).toBeTruthy()`
would therefore always be true. This meant that we not only didn't
wait for the elements to appear but also that our check never failed.

This has been fixed by awaiting the return. This has been handled
inside of a new test helper function to cut down on incidental noise.

The second bug was that, despite the type signaures in the webdriverio
type definitions, `app.client` and `app.client.$()` return different
types, whereas `app.client` returns a `Client` and `$()` returns a
"WebElement JSON Object" or something that _isn't_ the same as the docs
suggest.

This bug was an infinite hang when trying to wait on the elements to
appear since the returned value never resolved.

Now all calls to wait on an element are handled by
`app.client.waitForExist(selector, ...)` instead of using `$()`

Further we have introduced a new helper via the `__TEST__` macro
so that we can signal from inside the app specific events that we
want our tests to wait for. This is performed with `waitForEvent()`
* Appease the linter