Immediately handle trash/restore/delete operations
In this change we're clearing out the selected note immediate when performing
the trash/restore/delete-forever operations to avoid showing an unnecessary
interstitial Simplenote logo or the newly-rendered note contents.

In addition to clearing out the note state we are immediately triggering a
note filter operation to update the note list and avoid a similar awkward
in-between state.