Move search into a WebWorker
Previously we have had to deal with the inconvenient way in which search
operations occur in the same UI thread as the rest of the app. The inconvenient
part has been that while typing and when you want the app to be most responsive
- that is when it was the slowest due to the filtering happening in the note
list.

For many years this was such a problem and would make typing in the search
field too cumbersome so we created a debounce value to wait until you stopped
typing in order to update the note list. This debounce, carried out as a
separate data stream from the main app state, caused problems in that the
debounced search value and the app state could get out of sync with each other.
This led to even more confusion creating separate values for the text in the
search field _and_ a debounced search field value.

With recent refactorings to the app state we have been able to eliminate some
of the spurious data flows and get back to a single state value representing
the currently-typed value in the search field. We have separated the act of
updating the note list from the act of typing in a search query. With that
final necessary separate we are able to get to this patch.

In this patch we're recreating the search system and moving it into a
WebWorker, outside of the main UI thread. What this means is that we can now
update the filtering of the note list as often as we want and it won't
interfere with the search field or with typing. By moving the data into its own
asynchronous "search indexer" too we are able to make several trivial
optimizations which have resulted in improving the performance of searching by
several magnitudes. These mostly revolve around doing less work by being
reactive instead of repeating calculations on every keypress:

 - Instead of building a case-insensitive search RegExp, store the lower-cased
   version of each note in the search index and lower-case the search field and
use more basic string search functions.
 - Instead of parsing the search query every time we execute a search/filter
   operation, only parse it when it changes.
 - Instead of scanning the list of note tags on every search, build a `Set()`
   when we index the note which provides O(1) membership tests instead of O(N)
with the number of tags in a note.
 - Provide a guard against long-running searches by performing a "quick" search
   up until 10ms have passed and then queuing a longer full search if we hit
that threshold. (In all my testing I could not find a search that took more
than about 1ms).

In hind sight a few of the optimizations in this PR could have been merged into
`develop` directly without the WebWorker and because of the performance impact
we could probably get by without making search asynchronous. However, there is
still a clear benefit to pushing this into a WebWorker and out of the main
thread. As we continue to iterate on the quality of the search results this
will become more important.

Props to @belcherj and @codebykat for their hard work refactoring the app
state which has enabled this kind of work to be done.