Adjust webpack settings to inline webworker and avoid security issues (#1956)
When I deployed #1941 to staging.simplenote.com I discovered that the WebWorker wasn't loading. The reason was because it was being served from another domain than the website itself. This triggered content security policies in the browser.

With this change we're no longer loading an external script. Instead, webpack builds a BLOB containing the source of the worker and ships it inline with the rest of the application code, eliminating the external script load.

After re-deploying to staging I confirmed that this fixes the loading problem.