Refactor: Move system-tag updating into Redux and into middleware (#1928)
Previously we have been using a Simperium roundtrip in order to
update the pinned/markdown/publish status for a note. This involved
dispatching an action to set those system tags, waiting for a response,
and a second dispatch to update the app state by re-selecting the
changed note.

This involved a number of data races including some important ones
involved in updating the currently-selected note.

In this patch we've moved that update logic into Redux middleware
and adjusted the app reducers so that changes are reflected immediately
in the app and Simperium gets updated in the background.

If and when note updates fail or reflect different changes from the
remote side then we will update through the usual data flow for
receiving remote changes.

Because pinning, setting Markdown status, and publishing are all three
facets of the same interaction I have merged them into this one patch
and created action-creators that all collapse into the same action,
which is setting the status of a system tag.