Fix: Highlight search matches in Markdown preview views
Previously we have only been highlighting search matches in
the plaintext view. If you're viewing the preview of a note
you won't see any indication of where the matches were.

In this patch we're adding a new stage in the HTML rendering
pipeline for the preview that replaces search matches in the
note with the wrapping `<span>` that highlights it as a match.