Refactor: Move tag suggestions into Redux
Since we introduced tag suggestions in #1648 we have been recalculating
the list of tag suggestions on every render, at least twice (once to get
the count of tag suggestions in note-list and once in tag-suggestions).

In order to prevent needless re-rendering we implemented custom memoization
in the tag-suggestions component.

In this patch we're moving the tag suggestions logic into our search module
where the tag suggestions can update with the rest of the search results.
By doing this we'll leave one place to update them, couple them with the
filtered notes, and be able to remove our custom memoizatoin without losing
the benefits it brought us.