Note List: Replace custom height cache with one provided by react-virtualized (#1829)
Follows exploratory work in #1807

For a long time we have had our own custom note list height cache. This was
used to tell `react-virtualized` how tall each item in the list is so that it
can determine how many and which ones to render. That cache was complicated and
sensitive due to the caching and performance-oriented nature of the code.

Whether `CellMesurerCache` existed when we first put in `react-virtualized` I
don't know but it does do what we want it to do and with less guessing. It
keeps track of the heights for the rendered components and provides a clean API
to interface with the `AutoSizer`.

In this patch we're removing our custom height cache and instead relying on the
one provided by `react-virtualized`.

* special case the No Notes section so it retains its height

Co-authored-by: Kat Hagan <kat@codebykat.com>