Add keyboard shortcuts explainer (#1983)
Resolves #700
Resolves #798
Resolves #985
Resolves #1167
Resolves #1222

This application has had keyboard shortcuts since #603 but they have
been undocumented making them hard to discover and remember.

In this patch we're picking up the work started in #700 and adding
a keyboard shortcut overylay, opened through the Help menu or by
pressing `CmdOrCtrl` + `?`. This dialog shows a list of the
app-specific keyboard shortcuts and what they do.

* Review all keyboard shortcuts and fix browser issues

In this commit we've reviewed keyboard shortcut guidelines
and updated those in the app to avoid conflicting with normal
expectataions and OS shortcuts. This may cause some friction
but should improve the integration with the host operating
system and normal keyboard expectations.

* Add tooltips for shortcuts and link in browser for explainer

* Accept backspace or delete for trashing a note

* Show / key for keybinding help not ?

* Fix checklist keybinding by handling in app, not in Electron