Refactor: Remove previousIndex (#1989)
We have been tracking `previousIndex` when trashing, restoring, and deleting notes forever in order to properly find the next note to select after these operations. This process required tracking the `previuosIndex` through multiple contexts, some of which only sometimes tracked it. The result was a confusing system we couldn't touch without breaking.

In this patch we're eliminating the `previousIndex` altogether and moving the logic for selecting a note into the `note-list` where it holds the notion of a previous or a next note, a note above or a note below the current selection.

While this may ultimately be a pit-stop on the road to refactoring how `selectedNote` or `note` is stored in state it removes the state tracking that was so cumbersome and fragile.